Polish docs for v2.0.0b1 and switch OCM default to fp16
- Default ocm_inference_dtype: bf16 -> fp16 for broader GPU/MPS support;
  recommend fp32 for CPU inference.
- Backfill v2 items missing from the [Unreleased] changelog (AOI mode,
  show_progress, adaptive_tiling, tile_workers, GDAL network defaults,
  bounds-mode mask windowing, native-20m SCL, BIGTIFF=IF_SAFER, fp16
  default) and collapse the mid-cycle early_stop_missing_fraction
  rename + removal into a single v1.x->v2 line.
- Sidecar is written for both output_dir and output_path exports.
- README: correct tile_workers default to 8, broaden output_crs auto-pick
  to cover AOI mode, expand the mosaic_method perf tip to compare all
  four methods and note min/max_observations early stop.
- grid_id notebook: align "30m" comments with the actual resolution=40.
- Wide-area notebook: drop the _AWS suffix from the filename (the code
  defaults to MPC).

+++ s2mosaic/config.py
@@ -108,7 +108,7 @@ class MosaicRequest:
     scene_sort_fn: Optional[Callable[..., Any]] = None
     cloud_mask: str = CLOUD_MASK_OCM
     ocm_batch_size: int = 1
-    ocm_inference_dtype: str = "bf16"
+    ocm_inference_dtype: str = "fp16"
     tile_workers: Optional[int] = None
     adaptive_tiling: bool = True
     show_progress: bool = False
